requirements.txt
Flask==1.1.2
Werkzeug==1.0.1
matplotlib==3.4.1
numpy==1.20.2
requests==2.25.1
Pillow==8.2.0
tqdm==4.60.0